+++ src/scandeval/exceptions.py
@@ -43,3 +43,41 @@ class NoInternetConnection(Exception):
         """
         self.message = message
         super().__init__(self.message)
+
+
+class NaNValueInModelOutput(Exception):
+    """There is a NaN value in the model output."""
+
+    def __init__(self, message: str = "There is a NaN value in the model output."):
+        """Initialize the exception.
+
+        Args:
+            message:
+                The message to display.
+        """
+        self.message = message
+        super().__init__(self.message)
+
+
+class FlashAttentionNotInstalled(Exception):
+    """The `flash-attn` package has not been installed."""
+
+    def __init__(
+        self,
+        message: str = (
+            "The model you are trying to load requires Flash "
+            "Attention. To use Flash Attention, please install "
+            "the `flash-attn` package, which can be done by "
+            "running `pip install -U wheel && "
+            "FLASH_ATTENTION_SKIP_CUDA_BUILD=TRUE pip install "
+            "flash-attn --no-build-isolation`."
+        ),
+    ):
+        """Initialize the exception.
+
+        Args:
+            message:
+                The message to display.
+        """
+        self.message = message
+        super().__init__(self.message)