+++ src/scandeval/finetuning.py
@@ -24,7 +24,8 @@ from transformers.trainer import OptimizerNames
 
 from .callbacks import NeverLeaveProgressCallback
 from .config import BenchmarkConfig, DatasetConfig, ModelConfig
-from .exceptions import InvalidBenchmark
+from .enums import DataType
+from .exceptions import InvalidBenchmark, NaNValueInModelOutput
 from .model_loading import load_model
 from .protocols import Tokenizer
 from .utils import block_terminal_output, clear_memory, enforce_reproducibility
@@ -99,6 +100,14 @@ def finetune(
     """
     scores: dict[str, list[dict[str, float]]] = defaultdict(list)
 
+    using_cuda = benchmark_config.device == torch.device("cuda")
+    if using_cuda and torch.cuda.is_bf16_supported():
+        dtype = DataType.BF16
+    elif using_cuda:
+        dtype = DataType.FP16
+    else:
+        dtype = DataType.FP32
+
     bs: int = benchmark_config.batch_size
     for idx in itr:
         # Set variable that tracks whether we need to initialize new models in
@@ -133,6 +142,7 @@ def finetune(
                     benchmark_config=benchmark_config,
                     model_config=model_config,
                     iteration_idx=idx,
+                    dtype=dtype,
                     batch_size=bs,
                 )
 
@@ -165,6 +175,23 @@ def finetune(
                 logger.debug(f"Test scores for iteration {idx}: {itr_scores['test']}")
 
                 break
+
+            # NaN values can appear in the model output when using mixed precision, as
+            # the hidden states get overflowed. In this case we try to disable mixed
+            # precision and try again.
+            except NaNValueInModelOutput:
+                if dtype != DataType.FP32:
+                    dtype = DataType.FP32
+                    model_already_initialized = False
+                    logger.debug(
+                        "NaN value detected in model outputs while using mixed "
+                        "precision. Retrying with full fp32 precision."
+                    )
+                else:
+                    raise InvalidBenchmark(
+                        "NaN value detected in model outputs, even with mixed "
+                        "precision disabled."
+                    )
 
             except Exception as e:
                 if "CUDA" not in str(e) and "out of memory" not in str(e):
@@ -325,6 +352,13 @@ def finetune_single_iteration(
 
         return scores
 
+    except NaNValueInModelOutput as e:
+        del trainer
+        del model
+        del tokenizer
+        clear_memory()
+        raise e
+
     except (RuntimeError, ValueError, IndexError) as e:
         raise InvalidBenchmark(str(e))
 
@@ -333,6 +367,7 @@ def get_training_args(
     benchmark_config: BenchmarkConfig,
     model_config: ModelConfig,
     iteration_idx: int,
+    dtype: DataType,
     batch_size: int | None = None,
 ) -> TrainingArguments:
     """Get the training arguments for the current iteration.
@@ -345,6 +380,8 @@ def get_training_args(
         iteration_idx:
             The index of the current iteration. This is only used to generate a
             unique random seed for the current iteration.
+        dtype:
+            The data type to use for the model weights.
         batch_size:
             The batch size to use for the current iteration, or None if the batch size
             in the benchmark config should be used.
@@ -391,7 +428,8 @@ def get_training_args(
             load_best_model_at_end=True,
             optim=optimizer,
             seed=seed,
-            fp16=benchmark_config.device == torch.device("cuda"),
+            fp16=dtype == DataType.FP16,
+            bf16=dtype == DataType.BF16,
             disable_tqdm=not benchmark_config.progress_bar,
             ddp_find_unused_parameters=False,
         )