+++ src/scandeval/model_setups/utils.py
@@ -50,7 +50,6 @@ def setup_model_for_question_answering(model: PreTrainedModel) -> PreTrainedMode
         PreTrainedModel:
             The setup model.
     """
-
     # Get the models' token type embedding children, if they exist
     children = get_children_of_module(name="model", module=model)
 
@@ -154,11 +153,14 @@ def align_model_and_tokenizer(
     # Manually check that this model max length is valid for the model, and adjust
     # otherwise
     initial_max_length = tokenizer.model_max_length
+    first_non_special_token = next(
+        i for i in range(initial_max_length) if i not in tokenizer.all_special_ids
+    )
     for max_length in range(initial_max_length, 0, -1):
         tokenizer.model_max_length = max_length
-        dummy_inputs = torch.zeros(
-            1,
-            max_length,
+        dummy_inputs = torch.full(
+            size=(1, max_length),
+            fill_value=first_non_special_token,
             dtype=torch.long,
             device=model.device,
         )