+++ src/scandeval/question_answering.py
@@ -16,7 +16,8 @@ from .exceptions import InvalidBenchmark
 from .generation import extract_raw_predictions
 from .protocols import GenerativeModel, Tokenizer
 from .question_answering_trainer import QuestionAnsweringTrainer
-from .utils import get_special_token_metadata
+from .types import Labels, Predictions
+from .utils import get_special_token_metadata, raise_if_model_output_contains_nan_values
 
 logger = logging.getLogger(__package__)
 
@@ -133,7 +134,7 @@ class QuestionAnswering(BenchmarkDataset):
 
     def _compute_metrics(
         self,
-        model_outputs_and_labels: tuple[np.ndarray, np.ndarray],
+        model_outputs_and_labels: tuple[Predictions, Labels],
         id2label: list[str],
     ) -> dict[str, float]:
         """Compute the metrics needed for evaluation.
@@ -150,6 +151,8 @@ class QuestionAnswering(BenchmarkDataset):
             values.
         """
         model_outputs, labels = model_outputs_and_labels
+
+        raise_if_model_output_contains_nan_values(model_output=model_outputs)
 
         model_output_dtype = np.asarray(model_outputs).dtype
         if model_output_dtype in [np.float16, np.float32, np.float64]:
@@ -173,6 +176,7 @@ class QuestionAnswering(BenchmarkDataset):
                 if isinstance(scores, list):
                     scores = sum(scores) / len(scores)
                 results[cfg.name] = scores
+
         return results
 
     def _extract_few_shot_examples(