+++ src/scandeval/sequence_classification.py
@@ -20,7 +20,12 @@ from .config import DatasetConfig
 from .exceptions import InvalidBenchmark
 from .generation import extract_raw_predictions
 from .protocols import GenerativeModel, Tokenizer
-from .utils import GENERATIVE_MODEL_TASKS, get_special_token_metadata
+from .types import Labels, Predictions
+from .utils import (
+    GENERATIVE_MODEL_TASKS,
+    get_special_token_metadata,
+    raise_if_model_output_contains_nan_values,
+)
 
 logger = logging.getLogger(__package__)
 
@@ -125,7 +130,7 @@ class SequenceClassification(BenchmarkDataset):
 
     def _compute_metrics(
         self,
-        model_outputs_and_labels: tuple[np.ndarray, np.ndarray],
+        model_outputs_and_labels: tuple[Predictions, Labels],
         id2label: list[str],
     ) -> dict[str, float]:
         """Compute the metrics needed for evaluation.
@@ -142,6 +147,8 @@ class SequenceClassification(BenchmarkDataset):
             values.
         """
         model_outputs, labels = model_outputs_and_labels
+
+        raise_if_model_output_contains_nan_values(model_output=model_outputs)
 
         model_output_dtype = np.asarray(model_outputs).dtype
         if model_output_dtype in [np.float16, np.float32, np.float64]:
@@ -181,6 +188,7 @@ class SequenceClassification(BenchmarkDataset):
                 if isinstance(scores, list):
                     scores = sum(scores) / len(scores)
                 results[cfg.name] = scores
+
         return results
 
     def _extract_few_shot_examples(