+++ src/scandeval/utils.py
@@ -27,8 +27,10 @@ from transformers import logging as tf_logging
 
 from .config import Language
 from .enums import Framework
+from .exceptions import NaNValueInModelOutput
 from .languages import DA, NB, NN, NO, SV, get_all_languages
 from .protocols import GenerativeModel, Tokenizer
+from .types import Predictions
 
 logger = logging.getLogger(__package__)
 
@@ -155,6 +157,7 @@ def block_terminal_output():
     logging.getLogger("torch.distributed.distributed_c10d").setLevel(logging.ERROR)
     logging.getLogger("torch.nn.parallel.distributed").setLevel(logging.ERROR)
     logging.getLogger("vllm.engine.llm_engine").setLevel(logging.ERROR)
+    logging.getLogger("vllm.transformers_utils.tokenizer").setLevel(logging.ERROR)
 
     def init_vllm_logger(name: str):
         """Dummy function to initialise vLLM loggers with the ERROR level."""
@@ -543,3 +546,26 @@ def model_is_generative(model: PreTrainedModel | GenerativeModel) -> bool:
             return True
     except (NotImplementedError, TypeError):
         return False
+
+
+def raise_if_model_output_contains_nan_values(model_output: Predictions) -> None:
+    """Raise an exception if the model output contains NaN values.
+
+    Args:
+        model_output:
+            The model output to check.
+
+    Raises:
+        NaNValueInModelOutput:
+            If the model output contains NaN values.
+    """
+    if isinstance(model_output, np.ndarray):
+        if model_output.dtype == np.float32 and np.isnan(model_output).any():
+            raise NaNValueInModelOutput()
+    elif len(model_output) > 0:
+        if isinstance(model_output[0], str):
+            if any(x != x for x in model_output):
+                raise NaNValueInModelOutput()
+        elif len(model_output[0]) > 0:
+            if any(x != x for sublist in model_output for x in sublist):
+                raise NaNValueInModelOutput()